move IMatrixEvent into matrixtypes.ts
