+++ src/util.ts
@@ -31,33 +31,6 @@ export interface ICommandParameters {
     [index: string]: ICommandParameter;
 }
 
-export interface IMatrixEventContent {
-    body?: string;
-    info?: any; // tslint:disable-line no-any
-    name?: string;
-    topic?: string;
-    membership?: string;
-    msgtype?: string;
-    url?: string;
-    displayname?: string;
-    "m.relates_to"?: any; // tslint:disable-line no-any
-}
-
-export interface IMatrixEvent {
-    event_id: string;
-    state_key: string;
-    type: string;
-    sender: string;
-    room_id: string;
-    membership?: string;
-    avatar_url?: string;
-    displayname?: string;
-    redacts?: string;
-    content?: IMatrixEventContent;
-    unsigned?: any; // tslint:disable-line no-any
-    origin_server_ts?: number;
-}
-
 export class Util {
     /**
      * downloadFile - This function will take a URL and store the resulting data into
